+++ flag/service/helm/helm.go
@@ -2,9 +2,11 @@ package helm
 
 import (
 	"github.com/giantswarm/chart-operator/flag/service/helm/http"
+	"github.com/giantswarm/chart-operator/flag/service/helm/kubernetes"
 )
 
 type Helm struct {
 	HTTP            http.HTTP
+	Kubernetes      kubernetes.Kubernetes
 	TillerNamespace string
 }