+++ flag/service/helm/kubernetes/kubernetes.go
@@ -1,0 +1,5 @@
+package kubernetes
+
+type Kubernetes struct {
+	WaitTimeout string
+}