+++ service/controller/chart/chart.go
@@ -1,6 +1,8 @@
 package chart
 
 import (
+	"time"
+
 	"github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/k8sclient"
@@ -21,6 +23,7 @@ type Config struct {
 	K8sClient  k8sclient.Interface
 	Logger     micrologger.Logger
 
+	K8sWaitTimeout  time.Duration
 	TillerNamespace string
 }