+++ service/service.go
@@ -125,6 +125,7 @@ func New(config Config) (*Service, error) {
 			Logger:     config.Logger,
 			K8sClient:  k8sClient,
 
+			K8sWaitTimeout:  config.Viper.GetDuration(config.Flag.Service.Helm.Kubernetes.WaitTimeout),
 			TillerNamespace: config.Viper.GetString(config.Flag.Service.Helm.TillerNamespace),
 		}